requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# File Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Basic Text Processing
textdistance==4.6.0

# HTTP Client
requests==2.31.0

# Validation & Serialization
pydantic==2.5.1

# File Utilities
aiofiles==23.2.1